Hey, works already!  Checkpointing before I screw it up. Lots of the debugging has to go.  A lot of the problem was fsinit running more than once, and not converting filenames into strings (unicode is apparently bad)

+++ extfuse.py
@@ -50,7 +50,7 @@ def getParts(path):
     Return the slash-separated parts of a given path as a list
     """
     if path == os.sep:
-        return [[os.sep]]
+        return [os.sep]
     else:
         return path.split(os.sep)
 
@@ -73,13 +73,20 @@ class ExtFuse(Fuse):
     def __init__(self, *args, **kw):
         Fuse.__init__(self, *args, **kw)
 
+    already=False
+    @debugfunc
     def fsinit(self):
+        # Idempotent!
+        if self.already:
+            return
+        self.already=True
         # self.dbfile=os.tmpnam()
-        self.dbfile="EXTFS.db"
+        self.dbfile="/home/mark/EXTFS.db"
         try:
             os.unlink(self.dbfile)
         except OSError:
             pass
+        self.multithreaded=False # THIS can make it work!
         self.connection=sqlite.connect(self.dbfile)
         self.cursor=self.connection.cursor()
         self.cursor.execute(self.tablecommand)
@@ -91,24 +98,38 @@ class ExtFuse(Fuse):
         for fil in w:
             direc, name = fil.rsplit(os.path.sep,1)
             base, ext = os.path.splitext(name)
+            # self.DBG("Base({0}), Ext({1})".format(base,ext))
             base=base.replace('\\', '\\\\')
             base=base.replace("'", "\\'")
             ext=ext.replace('\\', '\\\\')
             ext=ext.replace("'", "\\'")
             if not ext:
-                ext='_'
+                ext='._'
+            ext=ext[1:]
             cmd=self.insertcommand.format(count, fil, base, ext)
             print cmd;
-            self.cursor.execute(cmd)
+            rv=self.cursor.execute(cmd)
             count+=1
         self.connection.commit()
+        self.connection.close() # ?
+        self.connection=sqlite.connect(self.dbfile)
+        self.cursor=self.connection.cursor()
+#        self.cursor.execute("SELECT * FROM files;")
+#        for l in self.cursor:
+#            print str(l)
+
+    @debugfunc
+    def fsdestroy(self):
+        self.cursor.close()
+        self.connection.close()
 
     @debugfunc
     def is_root(self, path=None, pathelts=None):
         if pathelts is None:
             pathelts=getParts(path)[1:]
         self.DBG("is_root ({0}), ({1})".format(str(path), str(pathelts)))
-        return path==os.sep or len(pathelts)==1
+        return (path==os.sep or len(pathelts)==0 or
+                pathelts == ['/'])
 
     # Depth is exactly two, after all.
     @debugfunc
@@ -118,7 +139,7 @@ class ExtFuse(Fuse):
         return len(pathelts) < 1
 
     @debugfunc
-    def getattr(self, path):
+    def _getattr(self, path):
         st=Stat()
         pe=getParts(path)
         st.st_mode = stat.S_IFDIR | 0555
@@ -139,11 +160,10 @@ class ExtFuse(Fuse):
             try:
                 self.DBG(query)
                 self.DBG("EJIIOFHSDKDH")
-                cc=sqlite.connect("EXTFS.db").cursor()
-                rv=cc.execute(query)
+                rv=self.cursor.execute(query)
                 self.DBG("AAAAAAAA")
                 self.DBG("exec returned {0}".format(str(rv)))
-                cnt=cc.fetchone()
+                cnt=self.cursor.fetchone()
                 self.DBG("Returned {0}".format(str(cnt)))
             except Exception as e:
                 self.DBG("Whoa, except: {0}".format(str(e)))
@@ -153,13 +173,37 @@ class ExtFuse(Fuse):
                 return -fuse.ENOENT
             return st
         else:
-            st.st_mode=stat.S_IFREG | 0444
+            # st.st_mode=stat.S_IFREG | 0444
+            st.st_mode=stat.S_IFLNK | 0777
             st.st_nlink=1
             st.st_size=0        # XXXXXXX
         return st
 
+
     @debugfunc
-    def readdir(self, path, offset):
+    def readlink(self, filename):
+        base, uniq=filename.rsplit('_',1)
+        if not uniq:
+            return -fuse.ENOENT
+        try:
+            query="SELECT fullpath FROM files WHERE _id={0}".format(int(uniq))
+        except ValueError:
+            return -fuse.ENOENT
+        self.DBG(query)
+        self.cursor.execute(query)
+        path=self.cursor.fetchone()
+        if not path or not path[0]:
+            return -fuse.ENOENT
+        return str(path[0])
+
+    def getattr(self, *args, **kwargs):
+        try:
+            return self._getattr(*args, **kwargs)
+        except Exception as e:
+            self.DBG("!!!, exception in getattr: {0}".format(str(e)))
+
+    @debugfunc
+    def _readdir(self, path, offset):
         dirents=['.', '..']
         pe=getParts(path)[1:]
         self.DBG("readdir pe: "+str(pe))
@@ -169,20 +213,32 @@ class ExtFuse(Fuse):
             self.DBG(query)
             self.cursor.execute(query)
             l=self.cursor.fetchall()
+            self.DBG("== {0}".format(str(l)))
             dirents.extend([x[0] for x in l])
             self.DBG("readdir returning {0}".format(str(dirents)))
             for r in dirents:
                 self.DBG("readdir yielding {0}".format(str(r)))
-                yield fuse.Direntry(r)
+                try:
+                    yield fuse.Direntry(str(r))
+                except Exception as e:
+                    self.DBG("Whoa, exception {0}".format(str(e)))
         elif len(pe)==1:
             query="SELECT newname FROM files WHERE ext='{0}';".format(escape_for_sql(pe[0]))
             self.DBG(query)
             self.cursor.execute(query)
             l=self.cursor.fetchone()
             while l:
-                yield fuse.Direntry(l[0])
+                self.DBG("File: {0}".format(str(l)))
+                yield fuse.Direntry(str(l[0]))
+                l=self.cursor.fetchone()
         else:
             raise StopIteration
+
+    def readdir(self, *args, **kwargs):
+        try:
+            return self._readdir(*args, **kwargs)
+        except Exception as e:
+            self.DBG("!!!, exception in readdir: {0}".format(str(e)))
 
     @debugfunc
     def mknod(self, path, mode, dev):
@@ -225,7 +281,10 @@ server=ExtFuse(version="%prog "+fuse.__version__,
                usage='', dash_s_do='setsingle')
 server.parser.add_option(mountopt='path', default='.')
 server.parse(errex=1, values=server)
-server.fsinit()
+try:
+    server.fsinit()
+except Exception as e:
+    print str(e)
 
 crs=server.connection.cursor()
 crs.execute("SELECT * FROM files;")
